packages/platform/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "artenic-ai-platform"
version = "0.6.0"
description = "Artenic AI Platform — Gateway, Registry, Orchestrator, Monitoring"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.12"
authors = [
    { name = "Artenic Cloud", email = "dev@artenic.cloud" },
]

dependencies = [
    "artenic-ai-sdk",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pydantic-settings>=2.0",
    "httpx>=0.27",
    "opentelemetry-api>=1.29",
    "opentelemetry-sdk>=1.29",
    "opentelemetry-instrumentation-fastapi>=0.50b",
    "opentelemetry-exporter-prometheus>=0.50b",
    "prometheus-client>=0.21",
    "mlflow>=2.17",
    "cryptography>=43.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
test = ["aiosqlite>=0.20", "httpx>=0.27"]
gcp = ["google-cloud-compute>=1.16", "google-cloud-storage>=2.14"]
aws = ["boto3>=1.34"]
azure = [
    "azure-mgmt-compute>=32.0",
    "azure-mgmt-network>=26.0",
    "azure-storage-blob>=12.19",
    "azure-identity>=1.15",
]
openstack = ["openstacksdk>=4.0"]
k8s = ["kubernetes>=28.0"]

[tool.hatch.build.targets.wheel]
packages = ["src/artenic_ai_platform"]
